CBT-21 fix couch_file startup
instead of using the ugly hack we used to handle eror on couch_file
init, reuse proc_lib:start_link/3 combined with proc_lib:init_ack/1.

While I am here clean the code so we retry opening the file if it failed
and just check its existence instead of trying to open it for read.