open and initialize a database.
cowdb can multiple stores. By defaul none is set. This commit allows us
to open a databasee from a file using the function `cowdb:open/{2,3}`
and pass to it an initialisation function:

    fun(Status, Db) ->
        ...

where Status can be:
- `init` (new database with no store)
- `current` : A database is opened with a version (by default 1). the
  current version on the disk is identical to the version passed to the
- `{upgrade, NewVersion, OldVersion}` when the version on the disk
  differs from the version given to the open function.

The status can be used to change a store definitiion, launch some
mmaintenance on the database, initialise it with some data or anything
else.

Example of usage:

    1> {ok, Pid} = cowdb:open("testdb", fun(Status, Db, _Args) ->
    1>  {ok, Store, Db1} = cowdb:init_store(Db, "test"),
    1>  {ok, Db1}
    1> end).
    {ok,<0.34.0>}
    2> cowdb:stores(Pid).
    ["test"]
    3>

Other design considerations:

all store roots are stored in a root btree which maintain the pointer to
them. The database header is really small and then has little chance to
be corrupted.