add transaction support.
This changes add support of transactions to cowdb.

A transaction received operations to execute as a list:

{add, StoreId, Obj} to add an object
{remove, StoreId, Key} to remove a value
{add_remove, StoreId, ToAdd, ToRemove} to add and remove multiple keys
and value at the same time
{fn, Func} a transaction function. A transaction function reveived the
db value like it was at the beginning of the transaction as an argument.
It's possible to pass arguments to it. A transaction function return a
list of operations and can wuery/maniuplate function. The list of
operations returned can also contain a function.

example of usage:

    1> {ok, Pid} = cowdb:open("testing.db",
    1> fun(St, Db) -> cowdb:open_store(Db, "test") end).
    {ok,<0.35.0>}
    2> cowdb:lookup(Pid, "test", [a,b]).
    [{ok,{a,1}},{ok,{b,2}}]
    3> cowdb:lookup(Pid, "test", [a,b,c]).
    [{ok,{a,1}},{ok,{b,2}},not_found]
    4> cowdb:transact(Pid, [{remove, "test", b}, {add, "test", {c,
3}}]).
    ok
    5> cowdb:lookup(Pid, "test", [a,b,c]).
    [{ok,{a,1}},not_found,{ok,{c,3}}]
    6> cowdb:get(Pid, "test", a).
    {ok,{a,1}}
    7> cowdb:transact(Pid, [{fn, fun(Db) -> [{add, "test", {d, 2}}]
end}]).
    ok
    8> cowdb:lookup(Pid, "test", [d]).
    [{ok,{d,2}}]