CBT-27 close add timeout support for transactions
this change allows you to pass a timeout to the transaction function to
not wait infinitly.

In the process cow_updater has been refactored, it is not anymore a
gen_server but a simple process started with proc_lib which is making
the code simpler to handle for other features.

Also the file reader is now started in the main db proces so it can
catch its exits

while I'am here close CBT-28.