refactor: make cowdb simple to use.
This change is a major refactoring. It remove the concept of having a
store / btree. Instead the database is usable as a simple KV db with a
reduce feature. The facilities offered by cbt to have multiples btree in
one file is used internally to handle the future feature of
snapshotting.

This change also make all commits really atomic which reduce the
rollback cost.

The new usage is the following:

    1> {ok, Pid} = cowdb:open("testing.db").
    {ok,<0.35.0>}
    2> cowdb:put(Pid, a).
    ok.
    3> cowdb:get(Pid, a).
    {ok,{a,1}}
    4> cowdb:lookup(Pid, [a, b]).
    [{ok,{a,1}},not_found]
    5> cowdb:put(Pid, b, 2).
    ok.
    6> cowdb:lookup(Pid, [a, b]).
    [{ok,{a,1}},{ok,{b,2}}]
    7> cowdb:lookup(Pid, [a, b, c, d]).
    [{ok,{a,1}},{ok,{b,2}},not_found,not_found]
    8> cowdb:transact(Pid, [
        {add, c, 2},
        {remove, b},
        {fn, fun(Db) ->
                    {ok, {a, V}} = cowdb:get(Db, a),
                    [{add, d, V}] end}]).
    ok
    9> cowdb:lookup(Pid, [a, b, c, d]).
[{ok,{a,1}},not_found,{ok,{c,2}},{ok,{d,1}}]
    10> cowdb:fold(Pid, fun(Got, Acc) -> {ok, [Got | Acc]} end, []).
    {ok,{[],[3]},[{d,1},{c,2},{a,1}]}