fix db initialisation
Old behaviour was bugged. Now, when no init function is given and the
database is created, an empty transaction is logged to save the initial
state. If an init function is given it now receive the old transaction
id and new transaction id so it can make choices basedon it.