add `cowdb:compact/1` - support compaction
We can now compact a cowdb database using the `cowdb:compact/1`
function.

Compaction is done using the following algorithm:

1. copy kVs frpm the id btree and copy value
2. commit the result in the new compact db
3. check if we are out of sync (if transaction id used to start the
compaction is less than the current transaction id)
4. if the compacted db is out of sync, then retry else switch the
compact file to the db file and remove the old one. We are doing it
gracefully by killing the reader after we renamed the compact file.

if the compaction is out of sync we retry it using this time the log
btree. We are doing from the last transction to the Transaction id of
the compactor. So we are only adding updates.